Delete package.json
Following the announcement made on #visual-music channel on Discord Wednesday 26th June 2019, I'm deleting the current app files so we can start again on a clean slate with better workflow and team involvement. Requirements to work on this repository are available in this announcement (pinned) and on the repository, please refer to it if necessary.